+++ src/lib/button/Button.svelte
@@ -1,4 +1,5 @@
 <script lang="ts">
+	import '#lib/internal/focus-ring.css'
 	import CircularProgress from '#lib/progress/CircularProgress.svelte'
 	import Ripple from '#lib/ripple/Ripple.svelte'
 	import Tooltip from '#lib/tooltip/Tooltip.svelte'
@@ -285,18 +286,7 @@
 		outline-color: var(--np-color-secondary);
 		outline-width: 3px;
 		outline-offset: 2px;
-		animation: focusAnimation 0.3s ease forwards;
-	}
-	@keyframes focusAnimation {
-		0% {
-			outline-width: 3px;
-		}
-		50% {
-			outline-width: 6px;
-		}
-		100% {
-			outline-width: 3px;
-		}
+		animation: focusAnimation var(--np-motion-expressive-slow-effects) forwards;
 	}
 	.text {
 		--np-ripple-hover-color: var(--np-text-button-label-text-color, var(--np-color-primary));