+++ src/lib/navigation-rail/NavigationRailItem.svelte
@@ -1,4 +1,5 @@
 <script lang="ts">
+	import '#lib/internal/focus-ring.css'
 	import Ripple from '#lib/ripple/Ripple.svelte'
 	import type { HTMLButtonAttributes } from 'svelte/elements'
 	import type { NavigationRailItemProps } from './types.ts'
@@ -60,18 +61,7 @@
 	}
 	@media (prefers-reduced-motion: no-preference) {
 		.np-navigation-action:focus-visible {
-			animation: focusAnimation 0.3s ease forwards;
-		}
-	}
-	@keyframes focusAnimation {
-		0% {
-			outline-width: 3px;
-		}
-		50% {
-			outline-width: 6px;
-		}
-		100% {
-			outline-width: 3px;
+			animation: focusAnimation var(--np-motion-expressive-slow-effects) forwards;
 		}
 	}
 	.np-navigation-action-icon {