+++ src/lib/radio/Radio.svelte
@@ -1,4 +1,5 @@
 <script lang="ts">
+	import '#lib/internal/focus-ring.css'
 	import Ripple from '#lib/ripple/Ripple.svelte'
 	import type { RadioProps } from './types.ts'
 
@@ -127,17 +128,6 @@
 		outline-color: var(--np-color-secondary);
 		outline-width: 3px;
 		outline-offset: 2px;
-		animation: focusAnimation 0.3s ease forwards;
-	}
-	@keyframes focusAnimation {
-		0% {
-			outline-width: 3px;
-		}
-		50% {
-			outline-width: 6px;
-		}
-		100% {
-			outline-width: 3px;
-		}
+		animation: focusAnimation var(--np-motion-expressive-slow-effects) forwards;
 	}
 </style>