+++ src/lib/select/Option.svelte
@@ -1,4 +1,5 @@
 <script lang="ts">
+	import '#lib/internal/focus-ring.css'
 	import Ripple from '#lib/ripple/Ripple.svelte'
 	import type { Snippet } from 'svelte'
 	import type { HTMLOptionAttributes } from 'svelte/elements'
@@ -58,22 +59,7 @@
 			outline-width: 3px;
 			outline-offset: -3px;
 			border-radius: var(--np-shape-corner-extra-small);
-			animation: focusAnimation 0.3s ease forwards;
-		}
-	}
-
-	@keyframes focusAnimation {
-		0% {
-			outline-offset: -3px;
-			outline-width: 3px;
-		}
-		50% {
-			outline-offset: -6px;
-			outline-width: 6px;
-		}
-		100% {
-			outline-offset: -3px;
-			outline-width: 3px;
+			animation: focusAnimationInset var(--np-motion-expressive-slow-effects) forwards;
 		}
 	}
 </style>